requirements_dev.txt
coverage
django-debug-toolbar
django-rosetta
pytest
pytest-cov
pytest-django
pytest-mock
django-rosetta
